# Source/onespace-appdiscovery-service/src/main/resources/swagger/AppDiscoveryFramework.yaml
openapi: 3.0.3
info:
  title: OneSpace Platform APIs
  description: |-
    __OneSpace Platform APIs__ enables OneSpace Portal to register, list & approve Philips  serviceability applications.
  version: 1.0.0
servers:
  - url: https://apps.onespace-{region}-{env}.com/api
...
paths:
...
  /Session/Cookie:
    post:
      tags:
        - Session Management APIs
      summary: Creates a new Session with OneSpace backend and returns a cookie.
      description: |
        Session is represented by a HTTP Cookie. The session cookie will be refreshed automatically when the original access token is about to expire.
      security: []
      requestBody:
        description: Provide the authorization code and redirect URI
        content:
          application/x-www-form-urlencoded:
            schema:
              type: string
              example: grant_type=authorization_code&code=ifggikAMzs_51TNwuxzRqOlrm8A&redirect_uri=http://localhost:8085/launchpad/
        required: true
      responses:
        '200':
          description: Successfully authenticated. The access & refresh token is returned as cookies. This needs to be included in subsequent requests.
          headers:
            Set-Cookie:
              schema:
                type: string
                example: onespacecookie=8279c45b58f7c29af1b48ec4019e6de6a4620c206f0ea53fbb494eb735bb6a9df64c5c365cd40747c640ef130cc1dad68c27bcaefa72d2f93bcc673dec142a0514ac802053482d299584678de9f5abfd9a7feacfb64d7c; Path=/; Max-Age=1799; Expires=Fri, 21 Jun 2024 11:53:06 GMT; Secure; SameSite=None
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '405':
          $ref: '#/components/responses/405'
        '406':
          $ref: '#/components/responses/406'
        '409':
          $ref: '#/components/responses/409'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '502':
          $ref: '#/components/responses/502'
        '503':
          $ref: '#/components/responses/503'
        '504':
          $ref: '#/components/responses/504'
  /Session/Token:
    post:
      tags:
        - Session Management APIs
      summary: Creates a new Session with OneSpace backend and returns bearer token.
      description: |
        Session is represented by a token.
      security: []
      requestBody:
        description: Provide the authorization code and redirect URI
        content:
          application/x-www-form-urlencoded:
            schema:
              type: string
              example: grant_type=authorization_code&code=ifggikAMzs_51TNwuxzRqOlrm8A&redirect_uri=http://localhost:8085/launchpad/
        required: true
      responses:
        '200':
          description: Successfully authenticated. The access & refresh token is returned. This needs to be included in subsequent requests.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '405':
          $ref: '#/components/responses/405'
        '406':
          $ref: '#/components/responses/406'
        '409':
          $ref: '#/components/responses/409'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '502':
          $ref: '#/components/responses/502'
        '503':
          $ref: '#/components/responses/503'
        '504':
          $ref: '#/components/responses/504'
  /Session/$Introspect:
    post:
      tags:
        - Session Management APIs
      summary: Returns the metadata associated with the Session Cookie / Bearer token
      responses:
        '200':
          description: Token metadata returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntrospectionResponse'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '405':
          $ref: '#/components/responses/405'
        '406':
          $ref: '#/components/responses/406'
        '409':
          $ref: '#/components/responses/409'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '502':
          $ref: '#/components/responses/502'
        '503':
          $ref: '#/components/responses/503'
        '504':
          $ref: '#/components/responses/504'
  /Session/$Exchange:
    post:
      tags:
        - Session Management APIs
      summary: Exchanges OneSpace issued Cookie with Access Token & Refresh Token
      description: |
        Enables a client registered in HSP IAM to present a HTTP cookie and get the access token, refresh token associated with it. Only cookies issued by OneSpace are allowed.
      security: []
      parameters:
        - in: header
          name: X-OneSpace-Cookie
          description: The encrypted OneSpace cookie value
          schema:
            type: string
          required: true
      requestBody:
        description: Details of client registered with HSP IAM
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientRequest'
        required: true
      responses:
        '200':
          description: Access & Refresh token returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '405':
          $ref: '#/components/responses/405'
        '406':
          $ref: '#/components/responses/406'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '502':
          $ref: '#/components/responses/502'
        '503':
          $ref: '#/components/responses/503'
        '504':
          $ref: '#/components/responses/504'
  /Session/$Logout:
    post:
      tags:
        - Session Management APIs
      summary: Logs out the user by revoking the access token & invalidating the session cookies.
      responses:
        '200':
          description: User logged out successfully
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '405':
          $ref: '#/components/responses/405'
        '406':
          $ref: '#/components/responses/406'
        '409':
          $ref: '#/components/responses/409'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '502':
          $ref: '#/components/responses/502'
        '503':
          $ref: '#/components/responses/503'
        '504':
          $ref: '#/components/responses/504'
  /Session/$Refresh:
    post:
      tags:
        - Session Management APIs
      summary: Refreshes the session by issuing a new Session Cookie.
      responses:
        '200':
          description: Session refreshed successfully
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '405':
          $ref: '#/components/responses/405'
        '406':
          $ref: '#/components/responses/406'
        '409':
          $ref: '#/components/responses/409'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '502':
          $ref: '#/components/responses/502'
        '503':
          $ref: '#/components/responses/503'
        '504':
          $ref: '#/components/responses/504'
...
  /Notification/Email:
    post:
      tags:
        - Notification APIs
      summary: Sends notification via email
      description: |
        A session cookie with __EPO_APP.LIST__ permission is required to execute this operation. If notification type is provided, a templated email based on the notification type is sent. Else a simple email with subject and message is sent.
      requestBody:
        description: Notification details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailNotification'
        required: true
      responses:
        '204':
          description: Email notification sent successfully
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '405':
          $ref: '#/components/responses/405'
        '406':
          $ref: '#/components/responses/406'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '502':
          $ref: '#/components/responses/502'
        '503':
          $ref: '#/components/responses/503'
        '504':
          $ref: '#/components/responses/504'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          description: Error Description message
          type: string
          example: Error while processing Request
        code:
          description: Error code
          type: string
          example: error_code
        source:
          description: Error Source
          type: string
          example: application_discovery_service
        category:
          description: Error Category
          type: string
          example: error_category
    BadRequestErrorResponse:
      type: object
      properties:
        message:
          description: Bad Request Error Description
          type: string
          example: Invalid input data provided.
        code:
          description: Error code
          type: string
          example: bad_request
        source:
          description: Error Source
          type: string
          example: application_discovery_service
        category:
          description: Error Category
          type: string
          example: client error
    BadRequestError:
      type: object
      properties:
        issue:
          description: Error for Bad Request scenerios
          type: array
          items:
            $ref: "#/components/schemas/BadRequestErrorResponse"
    UnAuthorizedErrorResponse:
      type: object
      properties:
        message:
          description: UnAuthorized Operation Error Description
          type: string
          example: UnAuthorized Access
        code:
          description: Error code
          type: string
          example: invalid_token
        source:
          description: Error Source
          type: string
          example: application_discovery_service
        category:
          description: Error Category
          type: string
          example: invalid_token
    UnAuthorizedRequestError:
      type: object
      properties:
        issue:
          description: Error for Bad Request scenerios
          type: array
          items:
            $ref: "#/components/schemas/UnAuthorizedErrorResponse"
...
    ResourceNotFoundErrorResponse:
      type: object
      properties:
        message:
          description: Error Message for Resource Not Found
          type: string
          example: Not Found
        code:
          description: Error code
          type: string
          example: resource_not_found
        source:
          description: Error Source
          type: string
          example: application_discovery_service
        category:
          description: Error Category
          type: string
          example: resource_not_found
    ResourceNotFoundError:
      type: object
      properties:
        issue:
          description: Error for Bad Request scenerios
          type: array
          items:
            $ref: "#/components/schemas/ResourceNotFoundErrorResponse"
    ConflictErrorResponse:
      type: object
      properties:
        message:
          description: Error Message for Conflicting Information
          type: string
          example: Resource already exists or Conflict with current state of resource.
        code:
          description: Error code
          type: string
          example: duplicate_key
        source:
          description: Error Source
          type: string
          example: application_discovery_service
        category:
          description: Error Category
          type: string
          example: duplicate_key
    ConflictError:
      type: object
      properties:
        issue:
          description: Error for Bad Request scenerios
          type: array
          items:
            $ref: "#/components/schemas/ConflictErrorResponse"
    InternalServerErrorResponse:
      type: object
      properties:
        message:
          description: Error Message for Internal Server Error
          type: string
          example: The server encountered an unexpected condition which prevented it from fulfilling the request.
        code:
          description: Error code
          type: string
          example: database_error
        source:
          description: Error Source
          type: string
          example: application_discovery_service
        category:
          description: Error Category
          type: string
          example: database_error
    InternalServerError:
      type: object
      properties:
        issue:
          description: Error for Bad Request scenerios
          type: array
          items:
            $ref: "#/components/schemas/InternalServerErrorResponse"
    Error:
      type: object
      properties:
        issue:
          description: A detailed human-readable message
          type: array
          items:
            $ref: "#/components/schemas/ErrorResponse"
    ClientRequest:
      type: object
      properties:
        clientId:
          description: ID of the client registered with HSP IAM
          type: string
          example: 05217113-b667-4c7d-9526-9cf166c9417a
        clientSecret:
          description: Secret of the client registered with HSP IAM
          type: string
          example: Secret@123
    TokenResponse:
      type: object
      properties:
        accessToken:
          description: Access token represented by the cookie
          type: string
          example: j-aNLUrx-FMz_7iglxlZKT5Km7Y
        refreshToken:
          description: Refresh token represented by the cookie
          type: string
          example: yrmO53jcEPCL6eVxxiczDW1ABN4
        expiresIn:
          description: The remaining time in seconds until this access token is valid.
          type: integer
          example: 1800
    IntrospectionResponse:
      type: object
      properties:
        active:
          type: boolean
          description: Boolean indicator of whether or not the presented token is
            currently active.
        scope:
          type: string
          description: A JSON string containing a space-separated list of scopes associated
            with this token, in the format described in Section 3.3 of OAuth 2.0 [RFC6749].
        client_id:
          type: string
          description: Client identifier for the OAuth 2.0 client that requested this
            token. ***Kindly note that the client identifier shall be returned in lower case,
            even if the OAuth 2.0 client has been created in mixed case fashion.***
        username:
          type: string
          description: Human-readable login ID for the resource owner who authorized
            this token.
        token_type:
          type: string
          description: Type of the token as defined in Section 5.1 of OAuth 2.0 [RFC6749].
        exp:
          type: integer
          description: Integer timestamp, measured in the number of seconds since
            January 1 1970 UTC, indicating when this token will expire, as defined
            in JWT [RFC7519].
          format: int64
        sub:
          type: string
          description: Subject of the token, as defined in JWT [RFC7519]. Usually
            a machine-readable identifier of the resource owner who authorized this
            token.
        iss:
          type: string
          description: String representing the issuer of this token, as defined in
            JWT.
        identity_type:
          type: string
          description: An enumeration that indicates the type of identity.
          enum:
            - device
            - user
            - service
            - client
        device_type:
          type: string
          description: If the identity is a device this string indicates what kind
            of device.
        organizations:
          $ref: '#/components/schemas/organizations'
      example:
        active: true
        scope: mail sn
        username: testuser@mycompany.com
        exp: 1555075957
        sub: b75c01f3-e0c0-482e-8b6b-2e6ffba71d6f
        iss: <am_host>/oauth2/access_token
        organizations:
          managingOrganization: d56f2d97-843d-491b-9499-b9a3ad029fe3
          organizationList:
            - organizationId: d56f2d97-843d-491b-9499-b9a3ad029fe3
              permissions:
                - GROUP.WRITE
                - ORGANIZATION.READ
              organizationName: MyCompany
              groups:
                - MyGroup
              roles:
                - MYROLE
            - organizationId: 2430bd92-e3e2-4295-aa37-d00b6fb47f2f
              permissions:
                - GROUP.WRITE
                - ORGANIZATION.READ
              organizationName: SomeCompany
              disabled: true
              groups:
                - MyGroup
              roles:
                - MYROLE
        client_id: MyClient
        token_type: Bearer
        identity_type: user
    organizations:
      type: object
      properties:
        managingOrganization:
          type: string
          description: HSDP UUID of the organization to which the identity account
            is registered.
        organizationList:
          type: array
          description: A list of organizations to which the user has been granted
            some form of access rights.  If `org_ctx` is passed, organizationList will contain only the details of the organization in context. Organization list will be empty if `org_ctx` is passed in the request with an
            organization where the subject of the token does not have any permissions
            assigned in the entire parent org structure.
          items:
            $ref: '#/components/schemas/organization'
...
    EmailNotification:
      type: object
      properties:
        to:
          description: The list of users for whom the email is to be sent
          type: array
          items:
            type: string
          minItems: 1
          example: [ "application.integrator@philips.com"]
        cc:
          description: The list of users for whom the email is to be cc'd
          type: array
          items:
            type: string
          example: [ "business.owner@philips.com", "market.approver@philips.com"]
        subject:
          description: The email subject
          type: string
          example: Application 'DoseWise Portal' registered
        message:
          description: The email content
          type: string
          example: Application 'DoseWise Portal' has been registered for Ultrasound business.
        type:
          type: string
          description: The type of email notification to be sent
          enum: [app-registered, app-business-owner-approved, app-solution-owner-approved]
          example: app-registered
        metadata:
          description: The metadata that will be used to populate the placeholders in email template
          type: object
          additionalProperties:
            type: string
          example:
            appName: DoseWise Portal
            appIcon: https://d1u65tovhg2ye9.cloudfront.net/icons/dosewiseportal
...
  responses:
...
    400:
      description: |
        **Bad Request**. Request is unparsable, syntactically incorrect, or violates 
        schema.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestError'
    401:
      description: |
        **Unauthorized**. Authorization failure due to missing or invalid session cookie.
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/UnAuthorizedRequestError'
    403:
      description: |
        **Forbidden**. Operation is not permitted based on the supplied authorization.
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/Error'
    404:
      description: |
        **Not Found**. Specified resource does not exist.
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/ResourceNotFoundError'
    405:
      description: |
        **Method Not Allowed**. The method specified for the request is not allowed.
    406:
      description: |
        **Not Acceptable**. The media type is invalid.
    409:
      description: |
        **Conflict**. Request server refused to create or update a resource that already exists or has conflicting information.
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/ConflictError'
    413:
      description: |
        **Request Entity Too Large**. The server is refusing to process a request because the 
        request entity is larger than the server is willing or able to process
    415:
      description: |
        **Unsupported Media Type**. The media type in the request is not supported.
    429:
      description: |
        **Too Many Requests**. The request quota has been filled. The server is throttling requests to prevent overload. 
    500:
      description: |
        **Internal Server Error**. The server encountered an unexpected condition which prevented 
        it from fulfilling the request.
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/InternalServerError'
    502:
      description: |
        **Bad Gateway**. The server received an invalid response from the upstream server.
    503:
      description: |
        **Service Unavailable**. The server is currently unable to handle the request due to a 
        temporary overloading or maintenance of the server.
    504:
      description: |
        **Gateway Timeout**. The server did not receive a timely response from the upstream 
        server specified it needed to access in attempting to complete the request 
...
  securitySchemes:
    onespaceCookieAuth:
      type: apiKey
      in: cookie
      name: onespacecookie
    bearerAuth:
      type: http
      scheme: bearer
security:
  - onespaceCookieAuth: []
  - bearerAuth: []